fix save config file bug ,Save_config_file

+++ bilang.cpp
@@ -5,7 +5,34 @@
 void Save_config_file(QString file_name);
 void Save_config_file(QString file_name)
 {
-    system("rm bilang.cfg");//文件名为 a.txt 可以用“C:/a.txt”之类
+    QString str_cfg1 = "bilang.cfg";
+    QString str_cfg2 = "bilang_2.cfg";
+    QString str_cfg3 = "bilang_3.cfg";
+    QString str_cfg4 = "bilang_4.cfg";
+    QString str_cfg5 = "bilang_5.cfg";
+    if(file_name == str_cfg1)
+    {
+        system("rm bilang.cfg");//文件名为 a.txt 可以用“C:/a.txt”之类
+    }
+    if(file_name == str_cfg2)
+    {
+        system("rm bilang_2.cfg");//文件名为 a.txt 可以用“C:/a.txt”之类
+    }
+    if(file_name == str_cfg3)
+    {
+        system("rm bilang_3.cfg");//文件名为 a.txt 可以用“C:/a.txt”之类
+    }
+    if(file_name == str_cfg4)
+    {
+        system("rm bilang_4.cfg");//文件名为 a.txt 可以用“C:/a.txt”之类
+    }
+    if(file_name == str_cfg5)
+    {
+        system("rm bilang_5.cfg");//文件名为 a.txt 可以用“C:/a.txt”之类
+    }
+
+
+
     QFile file(file_name);
     file.open(QIODevice::WriteOnly);
     for(int i=0;i<5;i++){
